提交

+++ F2/IO/FileIO.cs
@@ -9,21 +9,24 @@ namespace F
     {
         static FileIO()
         {
-            InstanceT.AddAssembly(typeof(F.FileIO).Assembly);
+            InstanceT.AddAssembly(typeof(FileIO).Assembly);
         }
+
         public static string GetFileSize(double length)
         {
-            var sizeConverters = new string[] { "B", "KB", "MB", "GB" };
+            var sizeConverters = new[] { "B", "KB", "MB", "GB" };
             var index = 0;
             while (length >= 1024f && index < sizeConverters.Length - 1)
             {
                 index++;
                 length /= 1024f;
             }
-            return length.ToString() + sizeConverters[index];
+
+            return length + sizeConverters[index];
         }
+
         /// <summary>
-        /// 获取文件大小
+        ///     获取文件大小
         /// </summary>
         /// <param name="path">文件路径</param>
         /// <returns></returns>
@@ -33,7 +36,7 @@ namespace F
         }
 
         /// <summary>
-        /// 指定文件写入流，文件不存在，创建文件
+        ///     指定文件写入流，文件不存在，创建文件
         /// </summary>
         /// <param name="path"></param>
         /// <param name="byteStream"></param>
@@ -42,45 +45,35 @@ namespace F
             CreateFile(path);
             File.WriteAllBytes(path, byteStream.Bytes);
         }
+
         public static void CreateFile(string path)
         {
-            if (!File.Exists(path))
-            {
-                File.Create(path).Dispose();
-            }
+            if (!File.Exists(path)) File.Create(path).Dispose();
         }
+
         public static void DeleteFile(string path)
         {
-            if (File.Exists(path))
-            {
-                File.Delete(path);
-            }
+            if (File.Exists(path)) File.Delete(path);
         }
+
         public static void CreateDirectory(string path)
         {
-            if (!Directory.Exists(path))
-            {
-                Directory.CreateDirectory(path);
-            }
+            if (!Directory.Exists(path)) Directory.CreateDirectory(path);
         }
+
         public static void DeleteDirectory(string path, bool recursive = true)
         {
-            if (Directory.Exists(path))
-            {
-                Directory.Delete(path, recursive);
-            }
+            if (Directory.Exists(path)) Directory.Delete(path, recursive);
         }
 
 
         public static void WriteBase64(string path, Dictionary<int, byte[]> v)
         {
             var streamWriter = new StreamWriter(path);
-            foreach (var item in v)
-            {
-                streamWriter.WriteLine($"{item.Key}|{Convert.ToBase64String(item.Value)}");
-            }
+            foreach (var item in v) streamWriter.WriteLine($"{item.Key}|{Convert.ToBase64String(item.Value)}");
             streamWriter.Close();
         }
+
         public static ByteStream WriteBytesDict<Tkey>(string path, Dictionary<Tkey, byte[]> v) where Tkey : unmanaged
         {
             var by = new ByteStream();
@@ -90,9 +83,11 @@ namespace F
                 by.Push(item.Key);
                 by.CopyFrom(item.Value, 0, item.Value.Length);
             }
+
             File.WriteAllBytes(path, by.Bytes);
             return by;
         }
+
         public static ByteStream WriteBytesDict(string path, Dictionary<string, byte[]> v)
         {
             var by = new ByteStream();
@@ -102,47 +97,97 @@ namespace F
                 by.Push(item.Key);
                 by.CopyFrom(item.Value, 0, item.Value.Length);
             }
+
             File.WriteAllBytes(path, by.Bytes);
             return by;
         }
 
-        public static ByteStream WriteBytesDict<Tkey>(string path, Dictionary<Tkey, IFSerializable> v, bool isWriteClassName = false) where Tkey : unmanaged
+        public static ByteStream WriteBytesDict<Tkey>(string path, Dictionary<Tkey, IFSerializable> v,
+            bool isWriteClassName = false) where Tkey : unmanaged
         {
             var s = new Serializable();
             s.PushLength(v.Count);
             foreach (var item in v)
             {
                 s.Push(item.Key);
-                if (isWriteClassName)
-                {
-                    s.Push(item.Value.GetType().FullName);
-                }
+                if (isWriteClassName) s.Push(item.Value.GetType().FullName);
                 item.Value.Serialization(s);
             }
+
             File.WriteAllBytes(path, s.ByteBuff);
             return s.ByteBuff;
         }
+
         /// <summary>
-        /// 写入序列化，支持多个不同类写入
+        ///     写入序列化，支持多个不同类写入
         /// </summary>
         /// <param name="path"></param>
         /// <param name="v"></param>
         /// <param name="isWriteClassName">是否写入类名</param>
-        public static ByteStream WriteBytesDict(string path, Dictionary<string, IFSerializable> v, bool isWriteClassName = false)
+        public static ByteStream WriteBytesDict(string path, Dictionary<string, IFSerializable> v,
+            bool isWriteClassName = false)
         {
             var s = new Serializable();
             s.PushLength(v.Count);
             foreach (var item in v)
             {
                 s.Push(item.Key);
-                if (isWriteClassName)
-                {
-                    s.Push(item.Value.GetType().FullName);
-                }
+                if (isWriteClassName) s.Push(item.Value.GetType().FullName);
                 item.Value.Serialization(s);
             }
+
             File.WriteAllBytes(path, s.ByteBuff);
             return s.ByteBuff;
+        }
+
+        /// <summary>
+        ///     写入序列化，支持多个不同类写入
+        /// </summary>
+        /// <param name="path"></param>
+        /// <param name="v"></param>
+        public static byte[] WriteBytesDict(Dictionary<string, IFSerializableKey> v)
+        {
+            using (var s = new SerializableKey
+                   {
+                       ByteBuff = new ByteStreamFixed(1024 * 1024)
+                   })
+            {
+                s.PushLength(v.Count);
+                foreach (var item in v)
+                {
+                    s.ByteBuff.Push(item.Key);
+                    var name = item.Value.GetType().FullName;
+                    s.ByteBuff.Push(name);
+                    s.PushSerializable(item.Value);
+                }
+
+                return s.ByteBuff;
+            }
+        }
+
+        /// <summary>
+        ///     读取序列化，支持多个不同类
+        /// </summary>
+        /// <typeparam name="TValue"></typeparam>
+        /// <param name="values"></param>
+        /// <returns></returns>
+        public static Dictionary<string, IFSerializableKey> ReadBytesDict(byte[] values)
+        {
+            using (var se = new SerializableKey(values))
+            {
+                var length = se.ReadLength();
+                var dict = new Dictionary<string, IFSerializableKey>(length);
+                while (length > 0)
+                {
+                    var key = se.Read();
+                    var toClass = InstanceT.CreateInstance<IFSerializableKey>(se.Read());
+                    se.ReadSerializable(ref toClass);
+                    dict.Add(key, toClass);
+                    length--;
+                }
+
+                return dict;
+            }
         }
 
         public static byte[] ReadBytes(string path)
@@ -150,15 +195,15 @@ namespace F
             return File.ReadAllBytes(path);
         }
 
-
         /// <summary>
-        /// 读取序列化，支持多个不同类
+        ///     读取序列化，支持多个不同类
         /// </summary>
         /// <typeparam name="TValue"></typeparam>
         /// <param name="values"></param>
         /// <param name="isReadClassName">是否读取类名</param>
         /// <returns></returns>
-        public static Dictionary<string, TValue> ReadBytesDict<TValue>(byte[] values, bool isReadClassName = false) where TValue : IFSerializable
+        public static Dictionary<string, TValue> ReadBytesDict<TValue>(byte[] values, bool isReadClassName = false)
+            where TValue : IFSerializable
         {
             var se = new Serializable(values);
             var length = se.ReadLength();
@@ -166,26 +211,34 @@ namespace F
             while (length > 0)
             {
                 var key = se.Read();
-                var toClass = isReadClassName ? InstanceT.CreateInstance<IFSerializable>(se.Read()) : InstanceT.CreateInstance<TValue>();
+                var toClass = isReadClassName
+                    ? InstanceT.CreateInstance<IFSerializable>(se.Read())
+                    : InstanceT.CreateInstance<TValue>();
                 toClass.Deserialization(se);
                 dict.Add(key, (TValue)toClass);
                 length--;
             }
+
             return dict;
         }
-        public static Dictionary<Tkey, TValue> ReadBytesDict<Tkey, TValue>(byte[] values, bool isReadClassName = false) where Tkey : unmanaged where TValue : IFSerializable
+
+        public static Dictionary<Tkey, TValue> ReadBytesDict<Tkey, TValue>(byte[] values, bool isReadClassName = false)
+            where Tkey : unmanaged where TValue : IFSerializable
         {
             var se = new Serializable(values);
-            var length =  se.ReadLength();
+            var length = se.ReadLength();
             var dict = new Dictionary<Tkey, TValue>(length);
             while (length > 0)
             {
                 var key = se.Read<Tkey>();
-                var toClass = isReadClassName ? InstanceT.CreateInstance<IFSerializable>(se.Read()) : InstanceT.CreateInstance<TValue>();
+                var toClass = isReadClassName
+                    ? InstanceT.CreateInstance<IFSerializable>(se.Read())
+                    : InstanceT.CreateInstance<TValue>();
                 toClass.Deserialization(se);
                 dict.Add(key, (TValue)toClass);
                 length--;
             }
+
             return dict;
         }
 
@@ -203,6 +256,7 @@ namespace F
                     return toClass;
                 }
             }
+
             throw new Exception($"ReadBase64 not found Id:{id}");
         }
 
@@ -218,21 +272,20 @@ namespace F
                 toClass.Deserialization(se);
                 dict.Add(Convert.ToInt16(bytesValue[0]), toClass);
             }
+
             return dict;
         }
+
         public static byte[] ReadBase64(string path, int id, string values)
         {
             var s = File.ReadAllLines(path);
             foreach (var item in s)
             {
                 var bytesValue = item.Split('|');
-                if (bytesValue[0] == id.ToString())
-                {
-                    return Convert.FromBase64String(bytesValue[1]);
-                }
+                if (bytesValue[0] == id.ToString()) return Convert.FromBase64String(bytesValue[1]);
             }
+
             throw new Exception($"ReadBase64 not found Id:{id}");
         }
-
     }
 }
